Initializer unit test fixed
